+++ README.md
@@ -1,1 +1,8 @@
 # AlertManager
+
+composer update
+
+install pdo_mysql
+
+run:
+    php app/AlertManager.php
